chore: split controller
Split AccountController to move visible signature methods to a dedicated
controller

Add specific header to identify the signer at public endpoints. This is
used when sign a document using email without account.
