+++ appinfo/routes.php
@@ -24,5 +24,6 @@ return array_merge_recursive(
 	include(__DIR__ . '/routes/routesPageController.php'),
 	include(__DIR__ . '/routes/routesRequestSignatureController.php'),
 	include(__DIR__ . '/routes/routesSettingsController.php'),
+	include(__DIR__ . '/routes/routesSignatureElementsController.php'),
 	include(__DIR__ . '/routes/routesSignFileController.php'),
 );
