+++ appinfo/routes/routesAccountController.php
@@ -17,12 +17,5 @@ return [
 		['name' => 'account#addFiles',                  'url' => '/api/{apiVersion}/account/files', 'verb' => 'POST', 'requirements' => $requirements],
 		['name' => 'account#deleteFile',                'url' => '/api/{apiVersion}/account/files', 'verb' => 'DELETE', 'requirements' => $requirements],
 		['name' => 'account#accountFileListToOwner',    'url' => '/api/{apiVersion}/account/files', 'verb' => 'GET', 'requirements' => $requirements],
-		['name' => 'account#accountFileListToApproval', 'url' => '/api/{apiVersion}/account/files/approval/list', 'verb' => 'GET', 'requirements' => $requirements],
-		['name' => 'account#createSignatureElement',    'url' => '/api/{apiVersion}/account/signature/elements', 'verb' => 'POST', 'requirements' => $requirements],
-		['name' => 'account#getSignatureElements',      'url' => '/api/{apiVersion}/account/signature/elements', 'verb' => 'GET', 'requirements' => $requirements],
-		['name' => 'account#getSignatureElementPreview','url' => '/api/{apiVersion}/account/signature/elements/preview/{nodeId}', 'verb' => 'GET', 'requirements' => $requirements],
-		['name' => 'account#getSignatureElement',       'url' => '/api/{apiVersion}/account/signature/elements/{nodeId}', 'verb' => 'GET', 'requirements' => $requirements],
-		['name' => 'account#patchSignatureElement',     'url' => '/api/{apiVersion}/account/signature/elements/{nodeId}', 'verb' => 'PATCH', 'requirements' => $requirements],
-		['name' => 'account#deleteSignatureElement',    'url' => '/api/{apiVersion}/account/signature/elements/{nodeId}', 'verb' => 'DELETE', 'requirements' => $requirements],
 	],
 ];
